Reorganized file/folder structure. Created register screen
