+++ app/components/forms/ErrorMessage/styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import colors from "../../config/colors";
+import colors from "../../../config/colors";
 
 const styles = StyleSheet.create({
   error: {
