+++ app/components/forms/SubmitButton/SubmitButton.js
@@ -2,7 +2,7 @@ import { View, Text } from "react-native";
 import React from "react";
 import { useFormikContext } from "formik";
 
-import AppButton from "../AppButton/AppButton";
+import AppButton from "../../AppButton/AppButton";
 
 export default function SubmitButton({ title }) {
   const { handleSubmit } = useFormikContext();
